- initial version with appveyor.yml configuration (#47)
- updated n++ plugin inteface files to https://github.com/npp-plugins/plugintemplate/releases/tag/v3.1
- adapt for vs2013, added jN.vcxproj
- adapt for x64

+++ common/AcceleratorTable.h
@@ -61,7 +61,7 @@ public:
 	~CAcceleratorTable(){
 		destroyTable();
 
-		for(int i=0, c=m_Accelerators.size(); i<c; i++){
+		for(size_t i=0, c=m_Accelerators.size(); i<c; i++){
 			delete m_Accelerators[i];
 		}
 	}
